requirements.txt
Django==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
mysqlclient==1.3.14
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2018.7
